+++ src/components/ProductsSliderCard.tsx
@@ -2,7 +2,7 @@ const ProductsSliderCard = ({ title, name, price, img }) => {
   return (
     <div className="w-[240px] h-[50px] flex flex-col justify-between group relative pb-[20px]">
       <div className="h-[300px] w-[240px]">
-        <img src={img} alt="" className="rounded-lg" />
+        <img src={img} alt="Electronic Products" className="rounded-lg" />
       
         <p className="absolute top-60 left-5 w-[210px] h-[44px] bg-white text-gray-600 text-sm py-2 rounded-lg drop-shadow-lg transition-opacity duration-300 ease-in-out opacity-0 group-hover:opacity-100">
           Quick view
