+++ src/components/SearchResult.tsx
@@ -6,12 +6,12 @@ const SearchResult = ({ search }: { search: string }) => {
   );
 
   return (
-    <div className="bg-gray-100 h-full pt-8 pb-8 overflow-y-auto">
+    <section className="bg-gray-100 h-full pt-8 pb-8 overflow-y-auto">
       {filterSearch.length > 0 ? (
         <div className="flex flex-col gap-2 size-4 w-full">
           {filterSearch.map((data) => (
             <p key={data.id} className="flex flex-row">
-              <img src={data.img} alt="" className="size-14" />
+              <img src={data.img} alt="products by search" className="size-14" />
               {data.name}$ {data.price}
             </p>
           ))}
@@ -19,7 +19,7 @@ const SearchResult = ({ search }: { search: string }) => {
       ) : (
         <p className="text-black">No results found</p>
       )}
-    </div>
+    </section>
   );
 };
