+++ src/components/static/ComapnyInfo.ts
@@ -1,3 +1,10 @@
+import Promotion1 from "@/assets/promotion-1.png";
+import Promotion2 from "@/assets/promotion-2.png";
+import Promotion3 from "@/assets/promotion-3.png";
+
+import productImg1 from "@/assets/p12.png";
+import productImg2 from "@/assets/p15.png";
+
 export const companyServices = [
     {
       id: 1,
@@ -92,5 +99,68 @@ export const companyServices = [
     {
       id: 9,
       Name: "Contact",
+    },
+  ];
+
+  export const features = [
+    {
+      id: 1,
+      title: "Up to 50% off on accessories",
+      img: Promotion1,
+    },
+    {
+      id: 2,
+      title: "Use automation to save energy",
+      img: Promotion2,
+    },
+    {
+      id: 3,
+      title: "Enjoy a greater level of home security",
+      img: Promotion3,
+    },
+  ];
+
+  export const bestSells = [
+    {
+      id: 1,
+      title: "Element Sync",
+      name: "Smart Home Speaker - Variant Image Set",
+      price: 67.7,
+      img: productImg1,
+    },
+    {
+      id: 2,
+      title: "Drifly",
+      name: "Indoor WiFi Smart Camera Pan & Tilt",
+      price: 89.0,
+      img: productImg2,
+    },
+    {
+      id: 3,
+      title: "Element Sync",
+      name: "Smart Home Speaker - Variant Image Set",
+      price: 67.7,
+      img: productImg1,
+    },
+    {
+      id: 4,
+      title: "DRIFLY",
+      name: "Indoor WiFi Smart Camera Pan & Tilt",
+      price: 89.0,
+      img: productImg2,
+    },
+    {
+      id: 5,
+      title: "Element Sync",
+      name: "Smart Home Speaker - Variant Image Set",
+      price: 67.7,
+      img: productImg1,
+    },
+    {
+      id: 6,
+      title: "Drifly",
+      name: "Indoor WiFi Smart Camera Pan & Tilt",
+      price: 89.0,
+      img: productImg2,
     },
   ];
